+++ protocols/gossipsub/CHANGELOG.md
@@ -12,6 +12,10 @@
 - Fix mesh not being constructed even when not adding any peer.
   See [PR 6100](https://github.com/libp2p/rust-libp2p/pull/6100)
 
+- Fix `unsubscribe_backoff` expecting number of seconds instead of `Duration`
+  See [PR 6124](https://github.com/libp2p/rust-libp2p/pull/6124)
+
+
 ## 0.49.0
 
 - Feature gate metrics related code. This changes some `Behaviour` constructor methods.