+++ protocols/gossipsub/CHANGELOG.md
@@ -1,7 +1,16 @@
 ## 0.50.0
+- Remove peer penalty for duplicate messages.
+  See [PR 6112](https://github.com/libp2p/rust-libp2p/pull/6112)
 
 - Remove `Rpc` from the public API.
   See [PR 6091](https://github.com/libp2p/rust-libp2p/pull/6091)
+
+## 0.49.2
+
+- Relax `Behaviour::with_metrics` requirements, do not require DataTransform and TopicSubscriptionFilter to also impl Default
+  See [PR 6097](https://github.com/libp2p/rust-libp2p/pull/6097)
+
+## 0.49.1
 
 - Fix applying P3 and P6 Score penalties when their weight is zero
   See [PR 6097](https://github.com/libp2p/rust-libp2p/pull/6097)